+++ flowcraft/generator/components/reads_quality_control.py
@@ -433,3 +433,37 @@ class DownsampleFastq(Process):
         self.status_channels = [
             "downsample_fastq"
         ]
+
+class Porechop(Process):
+    """Porechop trims adapters from Oxford Nanopore reads.
+
+    This process is set with:
+
+        - ``input_type``: fastq
+        - ``output_type``: fastq
+        - ``ptype``: pre_assembly
+    """
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+        self.input_type = "fastq"
+        self.output_type = "fastq"
+
+        self.link_end.append({"link": "raw_long_reads", "alias": "raw_long_reads"})
+        self.link_start.append("long_reads")
+
+        self.params = {
+            "long_reads": {
+                "default": "null",
+                "description": "FASTQ or FASTA file of long reads"
+            },
+        }
+
+        self.directives = {
+            "porechop": {
+                "cpus": 4,
+                "container": "quay.io/biocontainers/porechop",
+                "version": "0.2.3_seqan2.1.1--py36h2d50403_3"
+            }
+        }